+++ EnhanceQoL/Settings/MapNavigation.lua
@@ -503,12 +503,28 @@ data = {
 						parentSection = mapExpandable,
 					},
 					{
+						var = "squareMinimapStatsTimeUseClassColor",
+						text = L["squareMinimapStatsUseClassColor"] or "Use class color",
+						func = function(value)
+							addon.db["squareMinimapStatsTimeUseClassColor"] = value and true or false
+							applySquareMinimapStatsNow(true)
+						end,
+						default = false,
+						sType = "checkbox",
+						parent = true,
+						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsTime"),
+						notify = "squareMinimapStatsTime",
+						parentSection = mapExpandable,
+					},
+					{
 						var = "squareMinimapStatsTimeColor",
 						text = L["squareMinimapStatsColor"] or "Text color",
 						parent = true,
 						default = false,
 						sType = "colorpicker",
-						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsTime"),
+						parentCheck = function()
+							return isSquareMinimapStatElementEnabled("squareMinimapStatsTime")() and addon.db and addon.db.squareMinimapStatsTimeUseClassColor ~= true
+						end,
 						callback = function() applySquareMinimapStatsNow(true) end,
 						parentSection = mapExpandable,
 					},
@@ -690,12 +706,28 @@ data = {
 						parentSection = mapExpandable,
 					},
 					{
+						var = "squareMinimapStatsFPSUseClassColor",
+						text = L["squareMinimapStatsUseClassColor"] or "Use class color",
+						func = function(value)
+							addon.db["squareMinimapStatsFPSUseClassColor"] = value and true or false
+							applySquareMinimapStatsNow(true)
+						end,
+						default = false,
+						sType = "checkbox",
+						parent = true,
+						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsFPS"),
+						notify = "squareMinimapStatsFPS",
+						parentSection = mapExpandable,
+					},
+					{
 						var = "squareMinimapStatsFPSColor",
 						text = L["squareMinimapStatsColor"] or "Text color",
 						parent = true,
 						default = false,
 						sType = "colorpicker",
-						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsFPS"),
+						parentCheck = function()
+							return isSquareMinimapStatElementEnabled("squareMinimapStatsFPS")() and addon.db and addon.db.squareMinimapStatsFPSUseClassColor ~= true
+						end,
 						callback = function() applySquareMinimapStatsNow(true) end,
 						parentSection = mapExpandable,
 					},
@@ -899,12 +931,28 @@ data = {
 						parentSection = mapExpandable,
 					},
 					{
+						var = "squareMinimapStatsLatencyUseClassColor",
+						text = L["squareMinimapStatsUseClassColor"] or "Use class color",
+						func = function(value)
+							addon.db["squareMinimapStatsLatencyUseClassColor"] = value and true or false
+							applySquareMinimapStatsNow(true)
+						end,
+						default = false,
+						sType = "checkbox",
+						parent = true,
+						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsLatency"),
+						notify = "squareMinimapStatsLatency",
+						parentSection = mapExpandable,
+					},
+					{
 						var = "squareMinimapStatsLatencyColor",
 						text = L["squareMinimapStatsColor"] or "Text color",
 						parent = true,
 						default = false,
 						sType = "colorpicker",
-						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsLatency"),
+						parentCheck = function()
+							return isSquareMinimapStatElementEnabled("squareMinimapStatsLatency")() and addon.db and addon.db.squareMinimapStatsLatencyUseClassColor ~= true
+						end,
 						callback = function() applySquareMinimapStatsNow(true) end,
 						parentSection = mapExpandable,
 					},
@@ -981,12 +1029,17 @@ data = {
 						text = L["squareMinimapStatsLocationUseZoneColor"] or "Use zone color",
 						func = function(value)
 							addon.db["squareMinimapStatsLocationUseZoneColor"] = value and true or false
+							if value then addon.db["squareMinimapStatsLocationUseClassColor"] = false end
 							applySquareMinimapStatsNow(true)
 						end,
 						default = true,
 						sType = "checkbox",
 						parent = true,
-						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsLocation"),
+						parentCheck = function()
+							return isSquareMinimapStatElementEnabled("squareMinimapStatsLocation")()
+								and addon.db
+								and addon.db.squareMinimapStatsLocationUseClassColor ~= true
+						end,
 						notify = "squareMinimapStatsLocation",
 						parentSection = mapExpandable,
 					},
@@ -1058,12 +1111,32 @@ data = {
 						parentSection = mapExpandable,
 					},
 					{
+						var = "squareMinimapStatsLocationUseClassColor",
+						text = L["squareMinimapStatsUseClassColor"] or "Use class color",
+						func = function(value)
+							addon.db["squareMinimapStatsLocationUseClassColor"] = value and true or false
+							if value then addon.db["squareMinimapStatsLocationUseZoneColor"] = false end
+							applySquareMinimapStatsNow(true)
+						end,
+						default = false,
+						sType = "checkbox",
+						parent = true,
+						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsLocation"),
+						notify = "squareMinimapStatsLocation",
+						parentSection = mapExpandable,
+					},
+					{
 						var = "squareMinimapStatsLocationColor",
 						text = L["squareMinimapStatsColor"] or "Text color",
 						parent = true,
 						default = false,
 						sType = "colorpicker",
-						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsLocation"),
+						parentCheck = function()
+							return isSquareMinimapStatElementEnabled("squareMinimapStatsLocation")()
+								and addon.db
+								and addon.db.squareMinimapStatsLocationUseClassColor ~= true
+								and addon.db.squareMinimapStatsLocationUseZoneColor ~= true
+						end,
 						callback = function() applySquareMinimapStatsNow(true) end,
 						parentSection = mapExpandable,
 					},
@@ -1204,12 +1277,30 @@ data = {
 						parentSection = mapExpandable,
 					},
 					{
+						var = "squareMinimapStatsCoordinatesUseClassColor",
+						text = L["squareMinimapStatsUseClassColor"] or "Use class color",
+						func = function(value)
+							addon.db["squareMinimapStatsCoordinatesUseClassColor"] = value and true or false
+							applySquareMinimapStatsNow(true)
+						end,
+						default = false,
+						sType = "checkbox",
+						parent = true,
+						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsCoordinates"),
+						notify = "squareMinimapStatsCoordinates",
+						parentSection = mapExpandable,
+					},
+					{
 						var = "squareMinimapStatsCoordinatesColor",
 						text = L["squareMinimapStatsColor"] or "Text color",
 						parent = true,
 						default = false,
 						sType = "colorpicker",
-						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsCoordinates"),
+						parentCheck = function()
+							return isSquareMinimapStatElementEnabled("squareMinimapStatsCoordinates")()
+								and addon.db
+								and addon.db.squareMinimapStatsCoordinatesUseClassColor ~= true
+						end,
 						callback = function() applySquareMinimapStatsNow(true) end,
 						parentSection = mapExpandable,
 					},
@@ -1283,6 +1374,19 @@ data = {
 						step = 1,
 						default = -3,
 						sType = "slider",
+						parent = true,
+						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsTrackingButton"),
+						parentSection = mapExpandable,
+					},
+					{
+						var = "squareMinimapStatsTrackingButtonShowBackground",
+						text = L["squareMinimapStatsTrackingButtonShowBackground"] or "Show border/background",
+						func = function(value)
+							addon.db["squareMinimapStatsTrackingButtonShowBackground"] = value and true or false
+							applySquareMinimapStatsNow(true)
+						end,
+						default = true,
+						sType = "checkbox",
 						parent = true,
 						parentCheck = isSquareMinimapStatElementEnabled("squareMinimapStatsTrackingButton"),
 						parentSection = mapExpandable,
@@ -2264,6 +2368,7 @@ local squareMinimapStatsDefaults = {
 	squareMinimapStatsTimeOffsetY = 17,
 	squareMinimapStatsTimeFontSize = 18,
 	squareMinimapStatsTimeColor = { r = 1, g = 1, b = 1, a = 1 },
+	squareMinimapStatsTimeUseClassColor = false,
 	squareMinimapStatsTimeDisplayMode = "server",
 	squareMinimapStatsTimeUse24Hour = true,
 	squareMinimapStatsTimeShowSeconds = false,
@@ -2273,6 +2378,7 @@ local squareMinimapStatsDefaults = {
 	squareMinimapStatsFPSOffsetY = 3,
 	squareMinimapStatsFPSFontSize = 12,
 	squareMinimapStatsFPSColor = { r = 1, g = 1, b = 1, a = 1 },
+	squareMinimapStatsFPSUseClassColor = false,
 	squareMinimapStatsFPSThresholdMedium = 30,
 	squareMinimapStatsFPSThresholdHigh = 60,
 	squareMinimapStatsFPSColorLow = { r = 1, g = 0, b = 0, a = 1 },
@@ -2285,6 +2391,7 @@ local squareMinimapStatsDefaults = {
 	squareMinimapStatsLatencyOffsetY = 3,
 	squareMinimapStatsLatencyFontSize = 12,
 	squareMinimapStatsLatencyColor = { r = 1, g = 1, b = 1, a = 1 },
+	squareMinimapStatsLatencyUseClassColor = false,
 	squareMinimapStatsLatencyMode = "max",
 	squareMinimapStatsLatencyThresholdLow = 50,
 	squareMinimapStatsLatencyThresholdMid = 150,
@@ -2298,6 +2405,7 @@ local squareMinimapStatsDefaults = {
 	squareMinimapStatsLocationOffsetY = -3,
 	squareMinimapStatsLocationFontSize = 12,
 	squareMinimapStatsLocationColor = { r = 1, g = 1, b = 1, a = 1 },
+	squareMinimapStatsLocationUseClassColor = false,
 	squareMinimapStatsLocationShowZone = true,
 	squareMinimapStatsLocationShowSubzone = false,
 	squareMinimapStatsLocationSubzoneBelowZone = false,
@@ -2308,6 +2416,7 @@ local squareMinimapStatsDefaults = {
 	squareMinimapStatsCoordinatesOffsetY = -17,
 	squareMinimapStatsCoordinatesFontSize = 12,
 	squareMinimapStatsCoordinatesColor = { r = 1, g = 1, b = 1, a = 1 },
+	squareMinimapStatsCoordinatesUseClassColor = false,
 	squareMinimapStatsCoordinatesHideInInstance = true,
 	squareMinimapStatsCoordinatesDecimals = 2,
 	squareMinimapStatsCoordinatesUpdateInterval = 0.2,
@@ -2315,6 +2424,7 @@ local squareMinimapStatsDefaults = {
 	squareMinimapStatsTrackingButtonAnchor = "TOPLEFT",
 	squareMinimapStatsTrackingButtonOffsetX = 3,
 	squareMinimapStatsTrackingButtonOffsetY = -3,
+	squareMinimapStatsTrackingButtonShowBackground = true,
 	squareMinimapStatsTrackingButtonScale = 1.0,
 }
 
@@ -2326,6 +2436,7 @@ local squareMinimapStatsConfig = {
 		offsetYKey = "squareMinimapStatsTimeOffsetY",
 		fontSizeKey = "squareMinimapStatsTimeFontSize",
 		colorKey = "squareMinimapStatsTimeColor",
+		useClassColorKey = "squareMinimapStatsTimeUseClassColor",
 		anchorPoint = "BOTTOMLEFT",
 	},
 	fps = {
@@ -2335,6 +2446,7 @@ local squareMinimapStatsConfig = {
 		offsetYKey = "squareMinimapStatsFPSOffsetY",
 		fontSizeKey = "squareMinimapStatsFPSFontSize",
 		colorKey = "squareMinimapStatsFPSColor",
+		useClassColorKey = "squareMinimapStatsFPSUseClassColor",
 		anchorPoint = "BOTTOMLEFT",
 	},
 	latency = {
@@ -2344,6 +2456,7 @@ local squareMinimapStatsConfig = {
 		offsetYKey = "squareMinimapStatsLatencyOffsetY",
 		fontSizeKey = "squareMinimapStatsLatencyFontSize",
 		colorKey = "squareMinimapStatsLatencyColor",
+		useClassColorKey = "squareMinimapStatsLatencyUseClassColor",
 		anchorPoint = "BOTTOMRIGHT",
 	},
 	location = {
@@ -2353,6 +2466,7 @@ local squareMinimapStatsConfig = {
 		offsetYKey = "squareMinimapStatsLocationOffsetY",
 		fontSizeKey = "squareMinimapStatsLocationFontSize",
 		colorKey = "squareMinimapStatsLocationColor",
+		useClassColorKey = "squareMinimapStatsLocationUseClassColor",
 		anchorPoint = "TOP",
 	},
 	coordinates = {
@@ -2362,6 +2476,7 @@ local squareMinimapStatsConfig = {
 		offsetYKey = "squareMinimapStatsCoordinatesOffsetY",
 		fontSizeKey = "squareMinimapStatsCoordinatesFontSize",
 		colorKey = "squareMinimapStatsCoordinatesColor",
+		useClassColorKey = "squareMinimapStatsCoordinatesUseClassColor",
 		anchorPoint = "TOP",
 	},
 }
@@ -2496,12 +2611,14 @@ function addon.functions.applySquareMinimapTrackingButton()
 	local x = tonumber(addon.db and addon.db.squareMinimapStatsTrackingButtonOffsetX) or squareMinimapStatsDefaults.squareMinimapStatsTrackingButtonOffsetX or 0
 	local y = tonumber(addon.db and addon.db.squareMinimapStatsTrackingButtonOffsetY) or squareMinimapStatsDefaults.squareMinimapStatsTrackingButtonOffsetY or 0
 	local scale = clamp(tonumber(addon.db and addon.db.squareMinimapStatsTrackingButtonScale) or squareMinimapStatsDefaults.squareMinimapStatsTrackingButtonScale or 1, 0.5, 2.0)
+	local showBackground = not addon.db or addon.db.squareMinimapStatsTrackingButtonShowBackground ~= false
 
 	if holder:GetParent() ~= Minimap then holder:SetParent(Minimap) end
 	holder:SetFrameLevel((Minimap:GetFrameLevel() or 2) + 20)
 	holder:ClearAllPoints()
 	holder:SetPoint(point, Minimap, point, x, y)
 	holder:SetScale(scale)
+	holder.Background:SetShown(showBackground)
 
 	if button:GetParent() ~= holder then button:SetParent(holder) end
 	button:ClearAllPoints()
@@ -2521,6 +2638,14 @@ local function getSquareMinimapStatsColor(colorKey)
 		clamp(tonumber(color.g) or fallback.g or 1, 0, 1),
 		clamp(tonumber(color.b) or fallback.b or 1, 0, 1),
 		clamp(tonumber(color.a) or fallback.a or 1, 0, 1)
+end
+
+local function getSquareMinimapPlayerClassColor()
+	local classToken = (addon.variables and addon.variables.unitClass) or (UnitClass and select(2, UnitClass("player"))) or nil
+	local colors = CUSTOM_CLASS_COLORS or RAID_CLASS_COLORS
+	local classColor = colors and classToken and colors[classToken] or nil
+	if not classColor then return 1, 1, 1 end
+	return clamp(tonumber(classColor.r) or 1, 0, 1), clamp(tonumber(classColor.g) or 1, 0, 1), clamp(tonumber(classColor.b) or 1, 0, 1)
 end
 
 local function getSquareMinimapStatsFontPath()
@@ -2995,12 +3120,24 @@ end
 local function buildSquareMinimapFPSText(frame, renderCfg)
 	local fps = math.floor((GetFramerate() or 0) + 0.5)
 	local bucket = getSquareMinimapFPSBucket(renderCfg, fps)
-	if frame and frame._eqolFPSValue == fps and frame._eqolFPSBucket == bucket then return frame._eqolPrimaryText or "" end
 	local r, g, b = getSquareMinimapFPSBucketColor(renderCfg, bucket)
+	if
+		frame
+		and frame._eqolFPSValue == fps
+		and frame._eqolFPSBucket == bucket
+		and frame._eqolFPSColorR == r
+		and frame._eqolFPSColorG == g
+		and frame._eqolFPSColorB == b
+	then
+		return frame._eqolPrimaryText or ""
+	end
 	local text = ("FPS %s"):format(colorizeSquareMinimapText(fps, r, g, b))
 	if frame then
 		frame._eqolFPSValue = fps
 		frame._eqolFPSBucket = bucket
+		frame._eqolFPSColorR = r
+		frame._eqolFPSColorG = g
+		frame._eqolFPSColorB = b
 	end
 	return text
 end
@@ -3028,6 +3165,9 @@ local function buildSquareMinimapLatencyTexts(frame, renderCfg)
 	local worldBucket = getSquareMinimapLatencyBucket(renderCfg, world)
 	local maxValue = math.max(home, world)
 	local maxBucket = getSquareMinimapLatencyBucket(renderCfg, maxValue)
+	local hr, hg, hb = getSquareMinimapLatencyBucketColor(renderCfg, homeBucket)
+	local wr, wg, wb = getSquareMinimapLatencyBucketColor(renderCfg, worldBucket)
+	local mr, mg, mb = getSquareMinimapLatencyBucketColor(renderCfg, maxBucket)
 
 	if
 		frame
@@ -3037,12 +3177,19 @@ local function buildSquareMinimapLatencyTexts(frame, renderCfg)
 		and frame._eqolLatencyHomeBucket == homeBucket
 		and frame._eqolLatencyWorldBucket == worldBucket
 		and frame._eqolLatencyMaxBucket == maxBucket
+		and frame._eqolLatencyHomeColorR == hr
+		and frame._eqolLatencyHomeColorG == hg
+		and frame._eqolLatencyHomeColorB == hb
+		and frame._eqolLatencyWorldColorR == wr
+		and frame._eqolLatencyWorldColorG == wg
+		and frame._eqolLatencyWorldColorB == wb
+		and frame._eqolLatencyMaxColorR == mr
+		and frame._eqolLatencyMaxColorG == mg
+		and frame._eqolLatencyMaxColorB == mb
 	then
 		return frame._eqolPrimaryText or "", frame._eqolSecondaryText or ""
 	end
 
-	local hr, hg, hb = getSquareMinimapLatencyBucketColor(renderCfg, homeBucket)
-	local wr, wg, wb = getSquareMinimapLatencyBucketColor(renderCfg, worldBucket)
 	local homeText = ("H %sms"):format(colorizeSquareMinimapText(home, hr, hg, hb))
 	local worldText = ("W %sms"):format(colorizeSquareMinimapText(world, wr, wg, wb))
 	local primaryText, secondaryText = "", ""
@@ -3054,7 +3201,6 @@ local function buildSquareMinimapLatencyTexts(frame, renderCfg)
 		primaryText = homeText
 		secondaryText = worldText
 	else
-		local mr, mg, mb = getSquareMinimapLatencyBucketColor(renderCfg, maxBucket)
 		primaryText = ("MS %s"):format(colorizeSquareMinimapText(maxValue, mr, mg, mb))
 	end
 
@@ -3065,6 +3211,15 @@ local function buildSquareMinimapLatencyTexts(frame, renderCfg)
 		frame._eqolLatencyHomeBucket = homeBucket
 		frame._eqolLatencyWorldBucket = worldBucket
 		frame._eqolLatencyMaxBucket = maxBucket
+		frame._eqolLatencyHomeColorR = hr
+		frame._eqolLatencyHomeColorG = hg
+		frame._eqolLatencyHomeColorB = hb
+		frame._eqolLatencyWorldColorR = wr
+		frame._eqolLatencyWorldColorG = wg
+		frame._eqolLatencyWorldColorB = wb
+		frame._eqolLatencyMaxColorR = mr
+		frame._eqolLatencyMaxColorG = mg
+		frame._eqolLatencyMaxColorB = mb
 	end
 	return primaryText, secondaryText
 end
@@ -3105,6 +3260,7 @@ local function getSquareMinimapStatRenderConfig(statKey)
 		g = g,
 		b = b,
 		a = a,
+		useClassColor = cfg.useClassColorKey and addon.db[cfg.useClassColorKey] == true or false,
 		timeDisplayMode = statKey == "time" and (addon.db.squareMinimapStatsTimeDisplayMode or "server") or nil,
 		timeUse24Hour = statKey == "time" and addon.db.squareMinimapStatsTimeUse24Hour ~= false or nil,
 		timeShowSeconds = statKey == "time" and addon.db.squareMinimapStatsTimeShowSeconds == true or nil,
@@ -3186,7 +3342,9 @@ local function updateSquareMinimapStat(statKey)
 	end
 
 	local r, g, b, a = renderCfg.r, renderCfg.g, renderCfg.b, renderCfg.a
-	if renderCfg.useZoneColor then
+	if renderCfg.useClassColor then
+		r, g, b = getSquareMinimapPlayerClassColor()
+	elseif renderCfg.useZoneColor then
 		r, g, b = getSquareMinimapStatsZoneColor()
 		a = 1
 	end